changes in villager so they can only drop off at proper locations.

+++ Assets/Scripts/RTSCore/Constructible.cs
@@ -63,7 +63,7 @@ public class Constructible : Obstacle, IFactioned
         }
     }
 
-    public void TryBuild(int count, Actor builder = null)
+    public void TryBuild(float count, Actor builder = null)
     {
         AttributeHandler.Heal(count, AttributeChangeCause.HEALED, builder.AttributeHandler);
     }
